TASK-1 removed test method
