--- log/output.log
@@ -1,1 +1,0 @@
-2018-12-03 | 20:53:40 | INFO  | MlHhApplication:30 | Start method |
